requirements.txt
fastapi
uvicorn
gunicorn
redis
faiss-cpu
sentence-transformers
python-dotenv
google-generativeai
numpy==1.24.4